requirements.txt
joblib==1.2.0
numpy==1.24.3
opencv_python==4.8.0.76
opencv_python_headless==4.8.1.78
pandas==2.0.1
Pillow==10.3.0
scikit_learn==1.2.2
streamlit==1.33.0
torch==2.0.1
torchvision==0.15.2
